+++ BussinessLogicLayer/Services/NotesBusiness.cs
@@ -17,7 +17,7 @@ namespace BusinessLogicLayer.Services
         {
             INotesRep=iNotesRep;
         }
-        public FundooNoteEntity AddNotes(NotesModel notesModel, long userID)
+        public FundooNoteEntity AddNotes(NotesModel notesModel, int userID)
         {
             try
             {
@@ -28,7 +28,7 @@ namespace BusinessLogicLayer.Services
                 throw new Exception(ex.Message);
             }
         }
-        public List<FundooNoteEntity> GetAllNotes(long userID)
+        public List<FundooNoteEntity> GetAllNotes(int userID)
         {
             try
             {
@@ -50,7 +50,7 @@ namespace BusinessLogicLayer.Services
                 throw new Exception(ex.Message);
             }
         }
-        public bool Pin_UnPin_Note(int NoteID, long userID)
+        public bool Pin_UnPin_Note(int NoteID, int userID)
         {
             try
             {
@@ -61,7 +61,7 @@ namespace BusinessLogicLayer.Services
                 throw new Exception(ex.Message);
             }
         }
-        public bool Archive(int NoteID, long userID)
+        public bool Archive(int NoteID, int userID)
         {
             try
             {
@@ -72,7 +72,7 @@ namespace BusinessLogicLayer.Services
                 throw new Exception(ex.Message);
             }
         }
-        public bool Trash(int NoteID, long userID)
+        public bool Trash(int NoteID, int userID)
 
         {
             try
@@ -84,7 +84,7 @@ namespace BusinessLogicLayer.Services
                 throw new Exception(ex.Message);
             }
         }
-        public bool DeleteNotes(int NoteID, long userID)
+        public bool DeleteNotes(int NoteID, int userID)
         {
             try
             {
@@ -95,7 +95,7 @@ namespace BusinessLogicLayer.Services
                 throw new Exception(ex.Message);
             }
         }
-        public FundooNoteEntity UpdateNotes(NotesModel updateModel, int NoteID, long userID)
+        public FundooNoteEntity UpdateNotes(NotesModel updateModel, int NoteID, int userID)
         {
             try
             {
@@ -106,7 +106,7 @@ namespace BusinessLogicLayer.Services
                 throw new Exception(ex.Message);
             }
         }
-        public FundooNoteEntity Colour(string colour, int NoteID, long userID)
+        public FundooNoteEntity Colour(string colour, int NoteID, int userID)
         {
             try
             {
@@ -117,7 +117,7 @@ namespace BusinessLogicLayer.Services
                 throw new Exception(ex.Message);
             }
         }
-        public string UploadImage(string path, int NoteID, long userID)
+        public string UploadImage(string path, int NoteID, int userID)
         {
             try
             {
@@ -128,7 +128,7 @@ namespace BusinessLogicLayer.Services
                 throw new Exception(e.Message);
             }
         }
-        public string UploadImageFormFile(IFormFile path, int NoteID, long userID)
+        public string UploadImageFormFile(IFormFile path, int NoteID, int userID)
         {
             try
             {
@@ -139,7 +139,7 @@ namespace BusinessLogicLayer.Services
                 throw new Exception(e.Message);
             }
         }
-        public FundooNoteEntity Remainder(DateTime remainder, int NoteID, long userID)
+        public FundooNoteEntity Remainder(DateTime remainder, int NoteID, int userID)
         {
             try
             {
