+++ RepositoryLayer/Interface/INotesRepository.cs
@@ -9,18 +9,18 @@ namespace RepositoryLayer.Interface
 {
     public interface INotesRepository
     {
-        public FundooNoteEntity AddNotes(NotesModel notesModel, long userID);
-        public List<FundooNoteEntity> GetAllNotes(long userID);
+        public FundooNoteEntity AddNotes(NotesModel notesModel, int userID);
+        public List<FundooNoteEntity> GetAllNotes(int userID);
         public List<FundooNoteEntity> Get_All_Notes();
-        public bool Pin_UnPin_Note(int NoteID, long userID);
-        public bool Archive(int NoteID, long userID);
-        public bool Trash(int NoteID, long userID);
-        public bool DeleteNotes(int NoteID, long userID);
-        public FundooNoteEntity UpdateNotes(NotesModel updateModel, int NoteID, long userID);
-        public FundooNoteEntity Colour(string colour, int NoteID, long userID);
-        public string UploadImage(string path, int NoteID, long userID);
-        public string UploadImageFormFile(IFormFile path, int NoteID, long userID);
-        public FundooNoteEntity Remainder(DateTime remainder, int NoteID, long userID);
+        public bool Pin_UnPin_Note(int NoteID, int userID);
+        public bool Archive(int NoteID, int userID);
+        public bool Trash(int NoteID, int userID);
+        public bool DeleteNotes(int NoteID, int userID);
+        public FundooNoteEntity UpdateNotes(NotesModel updateModel, int NoteID, int userID);
+        public FundooNoteEntity Colour(string colour, int NoteID, int userID);
+        public string UploadImage(string path, int NoteID, int userID);
+        public string UploadImageFormFile(IFormFile path, int NoteID, int userID);
+        public FundooNoteEntity Remainder(DateTime remainder, int NoteID, int userID);
 
     }
 }
