+++ RepositoryLayer/Services/NotesRepository.cs
@@ -26,7 +26,7 @@ namespace RepositoryLayer.Services
             this.fundooContext = fundooContext;
             this.configuration = configuration;
         }
-        public FundooNoteEntity AddNotes(NotesModel notesModel,long userID)
+        public FundooNoteEntity AddNotes(NotesModel notesModel,int userID)
         {
             try
             {
@@ -41,7 +41,7 @@ namespace RepositoryLayer.Services
                 fundooNoteEntity.IsTrash=notesModel.IsTrash;
                 fundooNoteEntity.CreatedAt= notesModel.CreatedAt;
                 fundooNoteEntity.UpdatedAt= notesModel.UpdatedAt;
-                fundooNoteEntity.UserID = (int)userID;
+                fundooNoteEntity.UserID = userID;
                 fundooContext.FundooNotes.Add(fundooNoteEntity);
                 fundooContext.SaveChanges();
                 return fundooNoteEntity;
@@ -51,7 +51,7 @@ namespace RepositoryLayer.Services
                 throw new Exception(ex.Message);
             }
         }
-        public List<FundooNoteEntity> GetAllNotes(long userID)
+        public List<FundooNoteEntity> GetAllNotes(int userID)
         {
             try
             {
@@ -75,7 +75,7 @@ namespace RepositoryLayer.Services
                 throw new Exception(ex.Message);
             }
         }
-        public bool Pin_UnPin_Note(int NoteID,long userID)
+        public bool Pin_UnPin_Note(int NoteID, int userID)
         {
             try
             {
@@ -98,7 +98,7 @@ namespace RepositoryLayer.Services
                 throw new Exception(ex.Message);
             }
         }
-        public bool Archive(int NoteID,long userID)
+        public bool Archive(int NoteID, int userID)
         {
             try
             {
@@ -122,7 +122,7 @@ namespace RepositoryLayer.Services
                 throw new Exception(ex.Message);
             }
         }
-        public bool Trash(int NoteID,long userID)
+        public bool Trash(int NoteID, int userID)
         {
             try
             {
@@ -146,7 +146,7 @@ namespace RepositoryLayer.Services
                 throw new Exception(ex.Message);
             }
         }
-        public bool DeleteNotes(int NoteID,long userID)
+        public bool DeleteNotes(int NoteID, int userID)
         {
             try
             {
@@ -182,7 +182,7 @@ namespace RepositoryLayer.Services
                 throw new Exception (ex.Message);
             }
         }
-        public FundooNoteEntity UpdateNotes(NotesModel updateModel,int NoteID,long userID)
+        public FundooNoteEntity UpdateNotes(NotesModel updateModel,int NoteID, int userID)
         {
             try
             {
@@ -216,7 +216,7 @@ namespace RepositoryLayer.Services
                 throw new Exception(ex.Message);
             }
         }
-        public FundooNoteEntity Colour(string colour,int NoteID,long userID)
+        public FundooNoteEntity Colour(string colour,int NoteID, int userID)
         {
             try
             {
@@ -245,7 +245,7 @@ namespace RepositoryLayer.Services
                 throw new Exception (ex.Message);
             }
         }
-        public string UploadImage(string path,int NoteID,long userID)
+        public string UploadImage(string path,int NoteID, int userID)
         {
             try
             {
@@ -277,7 +277,7 @@ namespace RepositoryLayer.Services
                 throw new Exception(ex.Message); 
             }
         }
-        public string UploadImageFormFile(IFormFile path, int NoteID, long userID)
+        public string UploadImageFormFile(IFormFile path, int NoteID, int userID)
         {
             try
             {
@@ -309,7 +309,7 @@ namespace RepositoryLayer.Services
                 throw new Exception(ex.Message);
             }
         }
-        public FundooNoteEntity Remainder(DateTime remainder,int NoteID,long userID)
+        public FundooNoteEntity Remainder(DateTime remainder,int NoteID, int userID)
         {
             try
             {
